+++ src/test/ui/non-fmt-panic.rs
@@ -13,14 +13,21 @@ fn main() {
     core::panic!("Hello {}"); //~ WARN panic message contains an unused formatting placeholder
     assert!(false, "{:03x} {test} bla");
     //~^ WARN panic message contains unused formatting placeholders
+    assert!(false, S);
+    //~^ WARN panic message is not a string literal
     debug_assert!(false, "{{}} bla"); //~ WARN panic message contains braces
-    panic!(C); // No warning (yet)
-    panic!(S); // No warning (yet)
+    panic!(C); //~ WARN panic message is not a string literal
+    panic!(S); //~ WARN panic message is not a string literal
+    std::panic!(123); //~ WARN panic message is not a string literal
+    core::panic!(&*"abc"); //~ WARN panic message is not a string literal
     panic!(concat!("{", "}")); //~ WARN panic message contains an unused formatting placeholder
     panic!(concat!("{", "{")); //~ WARN panic message contains braces
 
     fancy_panic::fancy_panic!("test {} 123");
     //~^ WARN panic message contains an unused formatting placeholder
+
+    fancy_panic::fancy_panic!(S);
+    //~^ WARN panic message is not a string literal
 
     // Check that the lint only triggers for std::panic and core::panic,
     // not any panic macro:
@@ -28,4 +35,5 @@ fn main() {
         ($e:expr) => ();
     }
     panic!("{}"); // OK
+    panic!(S); // OK
 }
